requirements.txt
# UI
streamlit 

nest-asyncio
# PDF Processing Libraries
pypdf
pymupdf
pdfplumber
PyPDF2<3.0
camelot-py[cv]
Ghostscript

docling # IBM's Opensource
markitdown # Microsoft's Opensource
llama-cloud-services #llama-parse related 

# Computer Vision
opencv-python

# LLM related Libraries
ollama
openai
anthropic
langchain_ollama
langchain-groq
langchain_openai
langchain-anthropic
langchain-google-genai
langchain-unstructured
langchain_huggingface
langchain_community
tiktoken
google-genai

# Vector Store and Embeddings
faiss-cpu
sentence_transformers

# AWS Libraries
boto3
amazon-textract-caller>=0.2.0

# Utilities
python-dotenv


